% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_csv.R
\name{additor_csv}
\alias{additor_csv}
\title{Write a raw data table}
\usage{
additor_csv(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data to}
}
\value{
returns invisible\code{TRUE}
}
\description{
Add new data to the database. If a table exist, append it, if not, create a
new table.
}
